audio: volume: move volume header to volume folder
move them from include path to volume folder.
