+++ src/audio/volume/volume_hifi3.c
@@ -19,7 +19,7 @@
 
 LOG_MODULE_DECLARE(volume_hifi3, CONFIG_SOF_LOG_LEVEL);
 
-#include <sof/audio/volume.h>
+#include "volume.h"
 
 #ifdef VOLUME_HIFI3
